+++ indigo_app/forms/__init__.py
@@ -1,0 +1,5 @@
+from .documents import *
+from .places import *
+from .tasks import *
+from .users import *
+from .works import *
